Update 3017[LEARNING LOGS] Bill.py

+++ 3017[LEARNING LOGS] Bill.py
@@ -1,5 +1,5 @@
 """Bill."""
-price = float(input())
+price = int(input())
 service = 0
 
 if price < 500:
